requirements.txt
flask
bcrypt
requests
python-dotenv
PyJWT
flask-jwt-extended
